Working on Platform Landing Page

+++ MVC/CI_Platform_Web/CI_Platform.Models/Data/Timesheet.cs
@@ -13,7 +13,7 @@ public partial class Timesheet
 
     public TimeSpan? Time { get; set; }
 
-    public int? Action { get; set; }
+    public long? Action { get; set; }
 
     public DateTime DateVolunteered { get; set; }
